+++ src/containers/SideBar/SideBar.js
@@ -18,16 +18,13 @@
  */
 import React from 'react';
 import { observer, inject } from 'mobx-react';
-import { computed, observable, action } from 'mobx';
 
 import LayerToggle from '../../components/LayerToggle/LayerToggleSummer';
 
 import './SideBar.less';
+import EventsMenu from './EventsMenu';
 
-const dumpData = {
-  layer: ['cities'],
-  pins: ['persons', 'battle', 'document'],
-};
+const capitalizeFirstLetter = (s) => s.charAt(0).toUpperCase() + s.toLowerCase().slice(1);
 
 const Link = ({
   href, name, setRef, text
@@ -37,34 +34,24 @@ const Link = ({
   </a>
 );
 
+const LayerToggleButton = ({ click, name, styles = '' }) => (
+  <>
+    <div />
+    <LayerToggle
+      key={`layer_${click}`}
+      tooltip={capitalizeFirstLetter(name)}
+      place=''
+      extraClassName={styles}
+      checked
+      name={name}
+      click={() => click(name)}
+    />
+  </>
+);
+
 @inject('store')
 @observer
 class LayerControlWrapper extends React.Component {
-  @observable events = false;
-
-  @computed get eventsStyle() {
-    return {
-      justifyContent: 'flex-end',
-      width: '100%',
-      display: 'flex',
-      visibility: this.events === true ? 'initial' : 'hidden'
-    };
-  }
-
-  @computed get msg() {
-    return this.props.store.i18n.data;
-  }
-
-  nameToTooltip = (id) => this.msg.layerNames[id];
-
-  handleLayer = (data) => {
-    Object.keys(data.payload).map((cur) => {
-      this.props.store.flags[data.place].set(cur, data.payload[cur]);
-      this.props.store.analytics.metricHit(`main_${cur}`);
-      return false;
-    });
-  }
-
   handleZoom = (data) => {
     const out = Boolean(data.payload.minus);
     this.props.store.deck.zoomOut(out);
@@ -76,63 +63,13 @@ class LayerControlWrapper extends React.Component {
 
   click = (name) => this[name].click();
 
-  @action toggleEvents = () => {
-    this.events = !this.events;
-  }
-
   render() {
     return (
       <div className='side-bar__grid'>
-        <div style={this.eventsStyle}>
-          {Object.keys(dumpData).map((place) => (
-            dumpData[place].map((id) => (
-              <LayerToggle
-                key={`layer_${id}`}
-                tooltip={this.nameToTooltip(id)}
-                place={place}
-                checked={this.props.store.flags[place].list[id]}
-                name={id}
-                click={this.handleLayer}
-              />
-            ))))}
-        </div>
-        <LayerToggle
-          key='layer_events'
-          tooltip='Events'
-          place=''
-          checked={false}
-          name='events'
-          click={this.toggleEvents}
-        />
-        <div />
-        <LayerToggle
-          key='layer_admin'
-          tooltip='Admin'
-          place=''
-          // extraClassName='icon icon-magic-wand'
-          checked={false}
-          name='admin'
-          click={() => this.click('admin')}
-        />
-        <div />
-        <LayerToggle
-          key='layer_github'
-          tooltip='Github'
-          place=''
-          checked={false}
-          name='github'
-          click={() => this.click('github')}
-        />
-        <div />
-        <LayerToggle
-          key='layer_discord'
-          tooltip='Discord'
-          place=''
-          extraStyle={{ backgroundSize: '90% 90%' }}
-          checked={false}
-          name='discord'
-          click={() => this.click('discord')}
-        />
+        <EventsMenu />
+        <LayerToggleButton name='admin' click={this.click} />
+        <LayerToggleButton name='github' click={this.click} />
+        <LayerToggleButton name='discord' click={this.click} />
         <div style={{ visibility: 'hidden' }}>
           <Link href='/admin' name='admin' setRef={this.setRef} text='Link to admin interface' />
           <Link href='https://discord.gg/rN3uen5' name='discord' setRef={this.setRef} text='Link to Discord chat' />
